refactor(main): remove createDownload from Main
Main will now only export utils and other functions for direct consumption

BREAKING CHANGE: removed createDownload() function
